==> notebooks/1_Data_Cleaning.ipynb <==
# Spotify Vibe Shift: Data Cleaning & Preprocessing

## What I'm Doing Here

I used a basic ETL workflow - 
1. Load raw data
2. Transforming the dataset by handling the missing values, data types, duplicates
5. Export clean dataset for analysis


# Extracting the data file and importing libraries

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

df = pd.read_csv("spotify_songs.csv")

# Transforming the data

I am using these steps:

1. Remove duplicate entries
2. Verify data quality 
3. Handle missing values

df.head()

## Removing Duplicates

df = df.drop_duplicates()
df = df.reset_index(drop=True)

## Checking Data Types

df.info()

- All the data types are correct.

## Missing Values

missing = df.isnull().sum()
missing_pct = (missing / len(df))*100

missing_df = pd.DataFrame({'Missing':missing, 'Percentage':missing_pct})
print(missing_df[missing_df['Missing']>0])

df['track_name'] = df['track_name'].fillna("unknown")
df['track_artist'] = df['track_artist'].fillna("unknown")
df['track_album_name'] = df['track_album_name'].fillna("unknown")

# Export

df.to_csv('spotify.csv', index = False)
